src/requirements.txt
ipykernel
openai
pydantic-ai-slim
pydantic-monty
requests
